+++ Frends.SFTP.UploadFiles/CHANGELOG.md
@@ -6,7 +6,7 @@
 - Fixed issue when using invalid server fingerprint in MD5 string format throws wrong error message: Added more specific error messages.
 - Changed how MD5 string is handled. MD5 can now be given without ':' or '-' characters.
 - Fixed issue that Sha256 was only accepted in Base64 format: Added support for Sha256 in hex format.
-- Changed the used HostKeyAlgorithm by forcing to use ssh-rsa default was ed25519.
+- Added selector for host key algorithm which when enabled will force the task to use specific algorithm.
 - Added more tests for using server fingerprints.
 - Fixed issue where when using SourceOperation.Move the source file cannot be restored when exception occurs. 
 - Removed FileEncoding UTF-16 which was not implemented and threw exception if selected.